+++ pilz_robot_programming/CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog for package pilz_robot_programming
 
 Forthcoming
 -----------
+* Increase robot API version: 1.0.0 -> 2.0.0
 * Fix transformation of goal pose in robot API
 * Contributors: Pilz GmbH and Co. KG